finish styling of Vacation Row
