+++ src/container/VacationsOverview/VacationsOverview.js
@@ -3,15 +3,16 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import moment from 'moment';
 
-import { Resources, Utils } from '@data';
+import { Resources } from '@data';
 import { Page } from '@components';
 
-import type { ApiVacationsType, ApiVacationType } from '@data/Resources/ResourcesTypes';
+import type { ApiVacationsType } from '@data/Resources/ResourcesTypes';
 import type { AuthState } from '@data/Auth/AuthTypes';
 
-// import styles from './VacationsOverview.module.css';
+import { Vacation } from './components/Vacation';
+
+import styles from './VacationsOverview.module.css';
 
 const mapStateToProps = ({ isAuthenticated, user }: AuthState, { history }) => ({
     isAuthenticated,
@@ -20,36 +21,6 @@ const mapStateToProps = ({ isAuthenticated, user }: AuthState, { history }) => (
 });
 
 const mapDispatchToProps = () => ({});
-
-type VacationProps = {
-    vacation: ApiVacationType,
-};
-
-const Vacation = ({ vacation }: VacationProps) => {
-    const date = moment(vacation.day_date);
-    const fullDate = date.format('DD.MM.YYYY');
-    const shortDate = date.format('DD.MM');
-
-    const targetTime = Utils.formatDurationHoursToLocale(moment.duration(vacation.day_target_time));
-    return (
-        <fieldset className="">
-            <dl>
-                <dt className="visible-xs">Datum</dt>
-                <dd className="hidden-sm col-md-1">{fullDate}</dd>
-                <dd className="col-sm-1 visible-sm">{shortDate}</dd>
-
-                <dt className="visible-xs hidden-sm">Sollzeit</dt>
-                <dd className="col-sm-1 col-lg-1 hidden-sm">{targetTime}</dd>
-
-                <dt className="visible-xs">User</dt>
-                <dd className="col-sm-4 col-md-3 col-lg-2">{vacation.usr_firstname} {vacation.usr_lastname}</dd>
-
-                <dt className="visible-xs">Kommentar</dt>
-                <dd className="col-sm-6 col-md-7 col-lg-8">{vacation.per_comment}</dd>
-            </dl>
-        </fieldset>
-    );
-};
 
 export type VacationOverviewProps = {
     isAuthenticated: boolean,
@@ -86,7 +57,7 @@ export class VacationsOverviewContainer extends Component<VacationOverviewProps,
         if (!isAuthenticated) return null;
         return (
             <Page>
-                <fieldset className="hidden-xs">
+                <fieldset className={`hidden-xs ${styles.header}`}>
                     <dl>
                         <dt className="col-sm-2 col-md-1">Datum</dt>
                         <dt className="hidden-sm col-sm-1">Sollzeit</dt>
