+++ src/container/VacationsOverview/VacationsOverview.module.css
@@ -1,0 +1,3 @@
+.header {
+    padding: 20px 0 10px 0;
+}
