+++ src/container/VacationsOverview/components/Vacation/Vacation.md
@@ -1,0 +1,5 @@
+Basic Vacation:
+
+```jsx
+<Vacation>Vacation</Vacation>
+```
