+++ src/container/VacationsOverview/components/Vacation/Vacation.test.js
@@ -1,0 +1,10 @@
+/*
+import React from 'react';
+import { shallow, mount, render } from 'enzyme';
+import { Vacation } from './Vacation';
+*/
+describe('<Vacation />', () => {
+    it('Expect to have unit tests specified', () => {
+        expect(true).toEqual(true);
+    });
+});
