+++ src/container/VacationsOverview/components/Vacation/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { Vacation } from './Vacation';
+
